Moves CachePlayerDataPacket into core packet to match project structure
